Explicitly forbid "Infinity" and "NaN" in NetworkGraph #17
The JSON RFC explicitly forbids Infinity and NaN:
https://tools.ietf.org/html/rfc4627

"Numeric values that cannot be represented as sequences of digits (such
as Infinity and NaN) are not permitted."

NetJSON is based on the JSON RFC, so it makes sense to explicitly
forbid links with infinite weight.

It would also simplify a lot consumer implementations, which is
something I believe we have to strive for.

Closes #17